backend/requirements.txt
# E-BARMM Backend Dependencies
# FastAPI Application

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.3

# Authentication & Security
python-jose[cryptography]==3.2.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Object Storage (S3)
boto3==1.34.21
minio==7.2.3

# GIS
shapely==2.0.2
geopandas==0.14.2

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0
